Assignment 1
Cynthia Rose signs a loan with a Dutch bank. Cynthia receives 30 000 EUR right now (t = 0), 10 000 EUR one year later (t = 1) and 15 000 EUR another year later (t = 2). This loan should be paid back with 25 yearly payments, starting at t = 3. The payments reduce every year with 5%. Thus, if the first payment is - say - 1 000 EUR then the next one is 950 EUR, the third one is 902.5 EUR, and so on. Cynthia and her bank agree that the interest rate is 2.5% for the first 5 years, 3.5% in the next 5 years and 4% in the remaining period of time. Calculate the yearly loan payments.